fix: scope human-attention dialogs and document entry dedupe
Restrict interactive osascript dialogs to OperatorAttention payloads so
worker skipped/PR-ready action_required stays lightweight. Under
local-token auth, omit bare Open Loop deep links and fall back to Open
Log. Document EntryKey permanent-dedupe trade-offs and split the oversized
runtime contract test surface.

Generated-By: looper 0.11.4 (runner=fixer, agent=grok-build)
